append works! woo hoo!
on to precise wait times and better recording methods

git-svn-id: https://neurorighter.googlecode.com/svn/branches/NeuroRighter_ClosedLoop@196 471701b0-e017-11dd-9759-139c9ab14fc6

+++ NeuroRighter_05/Stimulation/File2Stim3.cs
@@ -163,7 +163,7 @@ namespace NeuroRighter
 
             // Wait for space to open in the buffer
             samplessent = stimAnalogTask.Stream.TotalSamplesGeneratedPerChannel;
-            while ((stimulusbuffer.NumBuffLoadsCompleted-1) * BUFFSIZE - samplessent > BUFFSIZE)
+            while ((stimulusbuffer.NumBuffLoadsCompleted - 1) * (ulong)BUFFSIZE - (ulong)samplessent > (ulong)BUFFSIZE)
             {
                 samplessent = stimAnalogTask.Stream.TotalSamplesGeneratedPerChannel;
             }
